Update .gitignore and ESLint rules, and fix ExamenSo(and refactoe) cioeconomico component
