+++ .gitignore
@@ -107,5 +107,6 @@ nosubir.txt
 .backend/.env
 
 .env
+.vscode/
 
 package-lock.json
